chap10/awesome-movies/requirements.txt
streamlit
# weaviate-client
python-dotenv
pandas
openai 
faiss-cpu